refactored and moved test directories
